backend/requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.11
Werkzeug==0.16.0
google-api-python-client==1.7.
Flask-Caching==1.8.0
flask-restful==0.3.7
isodate=0.6.0
python-dateutil=2.8.1